Add some support for email annotations/tags
At least for now, the email annotations are only suppported as tags of
the form "category:tag".

Which categories and tags are allowed can be set in the configuration
file (the example configuration has been extended accordingly). This
information is also included in the JSON object returned by
get_annotation_hints under the key "email_tags".

In the API the tags for an email-address are treated as a new part of
the email status and can be retrieved and set with the same API
enpoints. The tags are passed along as a new key, "tags", in the JSON
object. See the doc-strings for some more information.

This commit requires a corresponding change in intelmq-certbund-contact
(commit 7b29fed93e) which adds the database table where the
annotations/tags are stored.