Include information about the default tags in annotation hints
For each category of tags, one may now be marked as the default one for
that category. The data structure returned by get_annotation_hints now
includes that, by adding a new attribute default_tag to the object
associated with the category. Its value is the name of the default tag
of that category and will be one of the tags in the tags attribute.