Order the tag categories according to category_order
Use the new column category_order in the category table to order the
categories in the result of get_annotation_hints. The value associated
with the tags key in that result is now a list of [category-name, tags]
pairs (the tags in those pairs have the same format as before).

Requires intelmq-certbund-contact commit 582ccaaf0ddc