Experiment to evaluate JSON Schema for validation
When creating a new organisation, the organsation data is now validated
using a JSON schema. When this validation fails, a
HTTP_UNPROCESSABLE_ENTITY response is sent to the client with details
about the failure which the client could use to highlight the
corresponding parts of the UI.

This code is primarily intended to demonstrate how JSON schema might be
used for this. The Schema ist not necessarily complete and perhaps even
incorrect.